fix: try to fix all tests
